Cleanup (#48)
* Clean up code

* Remove unused imports
* Remove unneccessary self references
* Update some formatting
* Separate extensions in files
* Localize a string
* Add more swiftlint rules
* Remove --strict from swiftlint build phase to not treat warnings as errors

* Only print in DEBUG mode

Use `releasePrint` to even print in release mode

* Change voteAverage and voteCount type to Double

Update unknown voteCount, voteAverage and runtime behaviour

* Add date to export json file name

* Format array

* Reset styling for UIBarButtonItem in storyboard

+++ CineasteTests/Helper/CoreDataHelper.swift
@@ -64,7 +64,7 @@ class CoreDataHelper {
                              releaseDate: Date,
                              runtime: Int16,
                              title: String,
-                             voteAverage: Float,
+                             voteAverage: Double,
                              watched: Bool,
                              watchedDate: Date,
                              listPosition: Int16) {
